ortho-ia/desktop-agent/requirements.txt
flask>=3.0
flask-cors>=4.0
pillow>=10.0
pyautogui>=0.9.54
pystray>=0.19
requests>=2.31
